src/requirements.txt
datetime
discord
requests
mysql-connector-python
zoneinfo
